86682. Unable to reproduce the bug locally poing DB connectons to TST. Int32 is bigger than Int16.

+++ src/ESF/ESFA.DC.Summarisation.ESF.Data/Providers/ESFILRProvider.cs
@@ -86,7 +86,7 @@ namespace ESFA.DC.Summarisation.ESF.Data.Providers
                                         CalendarYear = 2000 + (ld.CollectionYear / 100),
                                         CalendarMonth = 8,
                                         Value = ld.AttributeName != "DeliverableVolume" ? ld.Period_1 : 0,
-                                        Volume = ld.AttributeName.Equals("DeliverableVolume") ? Convert.ToInt16(ld.Period_1) : 0
+                                        Volume = ld.AttributeName.Equals("DeliverableVolume") ? Convert.ToInt32(ld.Period_1) : 0
                                     },
                                     new Period
                                     {
@@ -94,8 +94,8 @@ namespace ESFA.DC.Summarisation.ESF.Data.Providers
                                         CollectionYear = ld.CollectionYear,
                                         CalendarYear = 2000 + (ld.CollectionYear / 100),
                                         CalendarMonth = 9,
-                                       Value = ld.AttributeName != "DeliverableVolume" ? ld.Period_2 : 0,
-                                        Volume = ld.AttributeName == "DeliverableVolume" ? Convert.ToInt16(ld.Period_2) : 0
+                                        Value = ld.AttributeName != "DeliverableVolume" ? ld.Period_2 : 0,
+                                        Volume = ld.AttributeName == "DeliverableVolume" ? Convert.ToInt32(ld.Period_2) : 0
                                     },
                                     new Period
                                     {
@@ -104,7 +104,7 @@ namespace ESFA.DC.Summarisation.ESF.Data.Providers
                                         CalendarYear = 2000 + (ld.CollectionYear / 100),
                                         CalendarMonth = 10,
                                         Value = ld.AttributeName != "DeliverableVolume" ? ld.Period_3 : 0,
-                                        Volume = ld.AttributeName == "DeliverableVolume" ? Convert.ToInt16(ld.Period_3) : 0
+                                        Volume = ld.AttributeName == "DeliverableVolume" ? Convert.ToInt32(ld.Period_3) : 0
                                     },
                                     new Period
                                     {
@@ -113,7 +113,7 @@ namespace ESFA.DC.Summarisation.ESF.Data.Providers
                                         CalendarYear = 2000 + (ld.CollectionYear / 100),
                                         CalendarMonth = 11,
                                         Value = ld.AttributeName != "DeliverableVolume" ? ld.Period_4 : 0,
-                                        Volume = ld.AttributeName == "DeliverableVolume" ? Convert.ToInt16(ld.Period_4) : 0
+                                        Volume = ld.AttributeName == "DeliverableVolume" ? Convert.ToInt32(ld.Period_4) : 0
                                     },
                                     new Period
                                     {
@@ -122,7 +122,7 @@ namespace ESFA.DC.Summarisation.ESF.Data.Providers
                                         CalendarYear = 2000 + (ld.CollectionYear / 100),
                                         CalendarMonth = 12,
                                         Value = ld.AttributeName != "DeliverableVolume" ? ld.Period_5 : 0,
-                                        Volume = ld.AttributeName == "DeliverableVolume" ? Convert.ToInt16(ld.Period_5) : 0
+                                        Volume = ld.AttributeName == "DeliverableVolume" ? Convert.ToInt32(ld.Period_5) : 0
                                     },
                                     new Period
                                     {
@@ -131,7 +131,7 @@ namespace ESFA.DC.Summarisation.ESF.Data.Providers
                                         CalendarYear = 2000 + (ld.CollectionYear % 100),
                                         CalendarMonth = 1,
                                         Value = ld.AttributeName != "DeliverableVolume" ? ld.Period_6 : 0,
-                                        Volume = ld.AttributeName == "DeliverableVolume" ? Convert.ToInt16(ld.Period_6) : 0
+                                        Volume = ld.AttributeName == "DeliverableVolume" ? Convert.ToInt32(ld.Period_6) : 0
                                     },
                                     new Period
                                     {
@@ -140,7 +140,7 @@ namespace ESFA.DC.Summarisation.ESF.Data.Providers
                                         CalendarYear = 2000 + (ld.CollectionYear % 100),
                                         CalendarMonth = 2,
                                         Value = ld.AttributeName != "DeliverableVolume" ? ld.Period_7 : 0,
-                                        Volume = ld.AttributeName == "DeliverableVolume" ? Convert.ToInt16(ld.Period_7) : 0
+                                        Volume = ld.AttributeName == "DeliverableVolume" ? Convert.ToInt32(ld.Period_7) : 0
                                     },
                                     new Period
                                     {
@@ -149,7 +149,7 @@ namespace ESFA.DC.Summarisation.ESF.Data.Providers
                                         CalendarYear = 2000 + (ld.CollectionYear % 100),
                                         CalendarMonth = 3,
                                         Value = ld.AttributeName != "DeliverableVolume" ? ld.Period_8 : 0,
-                                        Volume = ld.AttributeName == "DeliverableVolume" ? Convert.ToInt16(ld.Period_8) : 0
+                                        Volume = ld.AttributeName == "DeliverableVolume" ? Convert.ToInt32(ld.Period_8) : 0
                                     },
                                     new Period
                                     {
@@ -158,7 +158,7 @@ namespace ESFA.DC.Summarisation.ESF.Data.Providers
                                         CalendarYear = 2000 + (ld.CollectionYear % 100),
                                         CalendarMonth = 4,
                                         Value = ld.AttributeName != "DeliverableVolume" ? ld.Period_9 : 0,
-                                        Volume = ld.AttributeName == "DeliverableVolume" ? Convert.ToInt16(ld.Period_9) : 0
+                                        Volume = ld.AttributeName == "DeliverableVolume" ? Convert.ToInt32(ld.Period_9) : 0
                                     },
                                     new Period
                                     {
@@ -167,7 +167,7 @@ namespace ESFA.DC.Summarisation.ESF.Data.Providers
                                         CalendarYear = 2000 + (ld.CollectionYear % 100),
                                         CalendarMonth = 5,
                                         Value = ld.AttributeName != "DeliverableVolume" ? ld.Period_10 : 0,
-                                        Volume = ld.AttributeName == "DeliverableVolume" ? Convert.ToInt16(ld.Period_10) : 0
+                                        Volume = ld.AttributeName == "DeliverableVolume" ? Convert.ToInt32(ld.Period_10) : 0
                                     },
                                     new Period
                                     {
@@ -176,7 +176,7 @@ namespace ESFA.DC.Summarisation.ESF.Data.Providers
                                         CalendarYear = 2000 + (ld.CollectionYear % 100),
                                         CalendarMonth = 6,
                                         Value = ld.AttributeName != "DeliverableVolume" ? ld.Period_11 : 0,
-                                        Volume = ld.AttributeName == "DeliverableVolume" ? Convert.ToInt16(ld.Period_11) : 0
+                                        Volume = ld.AttributeName == "DeliverableVolume" ? Convert.ToInt32(ld.Period_11) : 0
                                     },
                                     new Period
                                     {
@@ -185,7 +185,7 @@ namespace ESFA.DC.Summarisation.ESF.Data.Providers
                                         CalendarYear = 2000 + (ld.CollectionYear % 100),
                                         CalendarMonth = 7,
                                         Value = ld.AttributeName != "DeliverableVolume" ? ld.Period_12 : 0,
-                                        Volume = ld.AttributeName == "DeliverableVolume" ? Convert.ToInt16(ld.Period_12) : 0
+                                        Volume = ld.AttributeName == "DeliverableVolume" ? Convert.ToInt32(ld.Period_12) : 0
                                     }
                                 }
                              }
